+++ src/objectpack/slave_object_pack/actions.py
@@ -1,13 +1,13 @@
 # coding:utf-8
-"""
+u"""
 Инструменарий для упрощённого создания ActionPack`ов для зависимых моделей
 """
 
 from m3_django_compat import ModelOptions
-import objectpack
+from objectpack.actions import ObjectPack
 
 
-class SlavePack(objectpack.ObjectPack):
+class SlavePack(ObjectPack):
     """
     "Ведомый" набор действий. Используется чаще всего для грида внутри
     окна редактирования объекта, отображающего объеты,