+++ src/objectpack/tree_object_pack/__init__.py
@@ -1,9 +1,4 @@
 # -*- coding: utf-8 -*-
-"""
-File: __init__.py
-Author: Rinat F Sabitov
-Description:
-"""
 
-from actions import TreeObjectPack, TreeObjectRowsAction
-from ui import BaseTreeListWindow, BaseTreeSelectWindow, BaseObjectTree
+from .actions import TreeObjectPack, TreeObjectRowsAction
+from .ui import BaseTreeListWindow, BaseTreeSelectWindow, BaseObjectTree